+++ app/buyers/views/digital_outcomes_and_specialists.py
@@ -2,12 +2,13 @@
 from __future__ import unicode_literals
 from app import data_api_client
 
-from flask import abort, render_template
-from ...helpers.buyers_helpers import get_framework_and_lot
-from ...main import main
+from flask import render_template
+
+from .. import dos
+from ..helpers.buyers_helpers import get_framework_and_lot
 
 
-@main.route('/buyers/frameworks/<framework_slug>/requirements/user-research-studios', methods=['GET'])
+@dos.route('/buyers/frameworks/<framework_slug>/requirements/user-research-studios', methods=['GET'])
 def studios_start_page(framework_slug):
     # Check framework is live and has the user-research-studios lot
     get_framework_and_lot(framework_slug, 'user-research-studios', data_api_client, status='live')
@@ -17,7 +18,7 @@ def studios_start_page(framework_slug):
     ), 200
 
 
-@main.route('/buyers/frameworks/<framework_slug>/requirements/<lot_slug>', methods=['GET'])
+@dos.route('/buyers/frameworks/<framework_slug>/requirements/<lot_slug>', methods=['GET'])
 def info_page_for_starting_a_brief(framework_slug, lot_slug):
     framework, lot = get_framework_and_lot(framework_slug, lot_slug, data_api_client,
                                            status='live', must_allow_brief=True)
