+++ app/main/__init__.py
@@ -4,6 +4,6 @@ main = Blueprint('main', __name__)
 
 from . import errors
 from .views import (
-    crown_hosting, digital_outcomes_and_specialists, digital_services_framework, g_cloud,
+    crown_hosting, digital_services_framework, g_cloud,
     login, marketplace, suppliers
 )
