+++ app/main/helpers/__init__.py
@@ -1,0 +1,9 @@
+import hashlib
+import base64
+
+
+def hash_email(email):
+    m = hashlib.sha256()
+    m.update(email.encode('utf-8'))
+
+    return base64.urlsafe_b64encode(m.digest())
