+++ app/main/helpers/shared_helpers.py
