+++ app/templates/buyers/create_brief_question.html
@@ -18,7 +18,7 @@
         "label": "Digital Marketplace"
       },
       {
-        "link": url_for("main.info_page_for_starting_a_brief", framework_slug=framework.slug, lot_slug=lot.slug),
+        "link": url_for("dos.info_page_for_starting_a_brief", framework_slug=framework.slug, lot_slug=lot.slug),
         "label": previous_title_text
       }
     ]
