+++ app/templates/index.html
@@ -27,22 +27,22 @@
     {% with
       items = [
         {
-          "link": url_for("main.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='digital-specialists'),
+          "link": url_for("dos.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='digital-specialists'),
           "title": "Find an individual specialist",
           "body": "eg a developer or user researcher",
         },
         {
-          "link": url_for("main.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='digital-outcomes'),
+          "link": url_for("dos.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='digital-outcomes'),
           "title": "Find a team to provide an outcome",
           "body": "eg a booking system or accessibility audit",
         },
         {
-          "link": url_for("main.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='user-research-participants'),
+          "link": url_for("dos.info_page_for_starting_a_brief", framework_slug=dos_slug, lot_slug='user-research-participants'),
           "title": "Find user research participants",
           "body": "eg people from a specific user group to test your service",
         },
         {
-          "link": url_for("main.studios_start_page", framework_slug=dos_slug),
+          "link": url_for("dos.studios_start_page", framework_slug=dos_slug),
           "title": "Find a user research lab",
           "body": "eg a room to conduct research sessions",
         },
